test base class bugfix
